+++ src/main/java/com/CuteNekoDragon/Core/utils/skills/skill/GlobalSkillData.java
@@ -1,0 +1,7 @@
+package com.CuteNekoDragon.Core.utils.skills.skill;
+
+public class GlobalSkillData {
+
+    public static double CURVE_BASE = 50.0;
+    public static double CURVE_EXPONENT = 1.6;
+}